Moved all from client to editor folder
