+++ core-web/libs/sdk/uve/src/internal.ts
@@ -1,5 +1,4 @@
 export * from './internal/index';
 
-export * from './lib/types/client/internal';
 export * from './lib/types/editor/internal';
 export * from './lib/types/events/internal';
