+++ core-web/libs/sdk/uve/src/lib/types/editor/internal.ts
@@ -1,3 +1,5 @@
+import { DotCMSUVEAction } from './public';
+
 /**
  * @description Custom client parameters for fetching data.
  */
@@ -18,4 +20,32 @@ export interface DotCMSReorderMenuConfig {
      * The depth of the menu levels to be reordered.
      */
     depth: number;
+}
+
+declare global {
+    interface Window {
+        dotCMSUVE: DotCMSUVE;
+    }
+}
+
+/**
+ * Post message props
+ *
+ * @export
+ * @template T
+ * @interface DotCMSUVEMessage
+ */
+export type DotCMSUVEMessage<T> = {
+    action: DotCMSUVEAction;
+    payload?: T;
+};
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+type DotCMSUVEFunction = (...args: any[]) => void;
+
+export interface DotCMSUVE {
+    editContentlet: DotCMSUVEFunction;
+    initInlineEditing: DotCMSUVEFunction;
+    reorderMenu: DotCMSUVEFunction;
+    lastScrollYPosition: number;
 }
