+++ core-web/libs/sdk/uve/src/types.ts
@@ -1,3 +1,2 @@
-export * from './lib/types/client/public';
 export * from './lib/types/editor/public';
 export * from './lib/types/events/public';
